Add PendingTransactionManager to manage pending transactions
- Introduced `PendingTransactionManager` class to handle the addition, tracking, and confirmation of pending transactions.
- Implemented methods for adding transactions, checking for pending transactions, and polling for confirmation status across different chains.
- Added safety cleanup for old transactions and ensured UI updates are performed on the main thread.
- Removed legacy pending transaction management code from `SendCryptoViewModel` to streamline the architecture.
